move all hash tests to single file
